refactor: 외부 API 호출 메서드를 AIService 서비스 클래스로 분리 (#146)
