+++ backend/src/main/java/com/tamnara/backend/news/service/NewsServiceImpl.java
@@ -1,8 +1,5 @@
 package com.tamnara.backend.news.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.tamnara.backend.bookmark.domain.Bookmark;
 import com.tamnara.backend.bookmark.repository.BookmarkRepository;
 import com.tamnara.backend.global.dto.WrappedDTO;
@@ -19,8 +16,6 @@ import com.tamnara.backend.news.dto.NewsCardDTO;
 import com.tamnara.backend.news.dto.NewsDetailDTO;
 import com.tamnara.backend.news.dto.StatisticsDTO;
 import com.tamnara.backend.news.dto.TimelineCardDTO;
-import com.tamnara.backend.news.dto.request.AINewsRequest;
-import com.tamnara.backend.news.dto.request.AITimelineMergeRequest;
 import com.tamnara.backend.news.dto.request.NewsCreateRequest;
 import com.tamnara.backend.news.dto.response.AINewsResponse;
 import com.tamnara.backend.news.dto.response.HotissueNewsListResponse;
@@ -41,24 +36,18 @@ import com.tamnara.backend.user.domain.Role;
 import com.tamnara.backend.user.domain.User;
 import com.tamnara.backend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.server.ResponseStatusException;
-import reactor.core.publisher.Mono;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
@@ -67,7 +56,7 @@ import java.util.concurrent.CompletionException;
 @RequiredArgsConstructor
 public class NewsServiceImpl implements NewsService {
 
-    private final WebClient aiWebClient;
+    private final AIService aiService;
     private final AsyncAIService asyncAiService;
 
     private final NewsRepository newsRepository;
@@ -230,7 +219,7 @@ public class NewsServiceImpl implements NewsService {
         try {
             LocalDate endAt = LocalDate.now();
             LocalDate startAt = endAt.minusDays(NEWS_CREATE_DAYS);
-            WrappedDTO<AINewsResponse> res = createAINews(req.getKeywords(), startAt, endAt);
+            WrappedDTO<AINewsResponse> res = aiService.createAINews(req.getKeywords(), startAt, endAt);
             aiNewsResponse = res.getData();
         } catch (AIException ex) {
             if (ex.getStatus() == HttpStatus.NOT_FOUND) {
@@ -240,7 +229,7 @@ public class NewsServiceImpl implements NewsService {
         }
 
         // 2. AI에 요청하여 타임라인 카드들을 병합한다.
-        List<TimelineCardDTO> timeline = mergeTimelineCards(aiNewsResponse.getTimeline());
+        List<TimelineCardDTO> timeline = aiService.mergeTimelineCards(aiNewsResponse.getTimeline());
 
         // 3. 뉴스의 여론 통계 생성 응답을 기다린다.
         WrappedDTO<StatisticsDTO> resStats = statsAsync.join();
@@ -359,7 +348,7 @@ public class NewsServiceImpl implements NewsService {
         try {
             LocalDate startAt = timelineCards.getFirst().getEndAt().plusDays(1);
             LocalDate endAt = LocalDate.now();
-            WrappedDTO<AINewsResponse> res = createAINews(keywords, startAt, endAt);
+            WrappedDTO<AINewsResponse> res = aiService.createAINews(keywords, startAt, endAt);
             aiNewsResponse = res.getData();
         } catch (AIException ex) {
             if (ex.getStatus() == HttpStatus.NOT_FOUND) {
@@ -370,7 +359,7 @@ public class NewsServiceImpl implements NewsService {
 
         // 4. 기존 타임라인 카드들과 합친 뒤, AI에게 요청하여 타임라인 카드들을 병합한다.
         oldTimeline.addAll(aiNewsResponse.getTimeline());
-        List<TimelineCardDTO> newTimeline = mergeTimelineCards(oldTimeline);
+        List<TimelineCardDTO> newTimeline = aiService.mergeTimelineCards(oldTimeline);
 
         // 2-2. 뉴스의 여론 통계 생성 응답을 기다린다.
         WrappedDTO<StatisticsDTO> resStats = statsAsync.join();
@@ -442,98 +431,6 @@ public class NewsServiceImpl implements NewsService {
         newsRepository.delete(news);
     }
 
-
-    /*
-        AI 통신용
-     */
-
-    private WrappedDTO<AINewsResponse> createAINews(List<String> keywords, LocalDate startAt, LocalDate endAt) {
-        AINewsRequest aiNewsRequest = new AINewsRequest(
-                keywords,
-                startAt,
-                endAt
-        );
-
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new JavaTimeModule());
-        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-
-        return aiWebClient.post()
-                .uri(TIMELINE_AI_ENDPOINT)
-                .bodyValue(aiNewsRequest)
-                .retrieve()
-                .onStatus(
-                        HttpStatusCode::isError,
-                        clientResponse -> clientResponse
-                                .bodyToMono(new ParameterizedTypeReference<WrappedDTO<AINewsResponse>>() {})
-                                .flatMap(errorBody -> Mono.error(new AIException(clientResponse.statusCode(), errorBody)))
-                )
-                .bodyToMono(new ParameterizedTypeReference<WrappedDTO<AINewsResponse>>() {})
-                .block();
-    }
-
-    private List<TimelineCardDTO> mergeTimelineCards(List<TimelineCardDTO> timeline) {
-        // 1. 1일카드 -> 1주카드
-        timeline = mergeAITimelineCards(timeline, TimelineCardType.DAY, 7);
-
-        // 2. 1주카드 -> 1달카드
-        timeline = mergeAITimelineCards(timeline, TimelineCardType.WEEK, 4);
-
-        // 3. 1달카드: 3개월 지남 -> 삭제
-        timeline.removeIf(tc -> (TimelineCardType.valueOf(tc.getDuration()) == TimelineCardType.MONTH)
-                && (tc.getStartAt().isBefore(LocalDate.now().minusMonths(3))));
-
-        return timeline;
-    }
-
-    private List<TimelineCardDTO> mergeAITimelineCards(List<TimelineCardDTO> timeline, TimelineCardType duration, Integer countNum) {
-        timeline.sort(Comparator.comparing(TimelineCardDTO::getStartAt));
-
-        List<TimelineCardDTO> mergedList = new ArrayList<>();
-        List<TimelineCardDTO> temp = new ArrayList<>();
-
-        int count = 0;
-
-        for (TimelineCardDTO tc : timeline) {
-            if (TimelineCardType.valueOf(tc.getDuration()) != duration) {
-                mergedList.add(tc);
-                continue;
-            }
-
-            temp.add(tc);
-            count++;
-
-            if (count == countNum) {
-                AITimelineMergeRequest mergeRequest = new AITimelineMergeRequest(temp);
-
-                WrappedDTO<TimelineCardDTO> merged = aiWebClient.post()
-                        .uri(MERGE_AI_ENDPOINT)
-                        .bodyValue(mergeRequest)
-                        .retrieve()
-                        .onStatus(
-                                HttpStatusCode::isError,
-                                clientResponse -> clientResponse
-                                        .bodyToMono(new ParameterizedTypeReference<WrappedDTO<TimelineCardDTO>>() {})
-                                        .flatMap(errorBody -> Mono.error(new AIException(clientResponse.statusCode(), errorBody)))
-                        )
-                        .bodyToMono(new ParameterizedTypeReference<WrappedDTO<TimelineCardDTO>>() {})
-                        .block();
-
-                mergedList.add(Objects.requireNonNull(merged).getData());
-
-                temp.clear();
-                count = 0;
-            }
-        }
-
-        mergedList.addAll(temp);
-        temp.clear();
-
-        timeline = mergedList;
-        timeline.sort(Comparator.comparing(TimelineCardDTO::getStartAt).reversed());
-
-        return timeline;
-    }
 
 
     /*
